textarea component tests
